Set up Transform as a generic action
