+++ src/CommandingOfficers/Meridian.java
@@ -4,23 +4,13 @@ import java.util.ArrayList;
 
 import CommandingOfficers.Modifiers.CODamageModifier;
 import CommandingOfficers.Modifiers.CODefenseModifier;
-import Engine.GameAction.WaitAction;
-import Engine.GameActionSet;
-import Engine.Path;
 import Engine.UnitActionType;
-import Engine.XYCoord;
 import Engine.Combat.BattleInstance.BattleParams;
-import Engine.GameEvents.GameEvent;
-import Engine.GameEvents.GameEventListener;
 import Engine.GameEvents.GameEventQueue;
 import Terrain.GameMap;
 import Terrain.MapMaster;
-import UI.MapView;
-import UI.Art.Animation.GameAnimation;
 import Units.Unit;
 import Units.UnitModel;
-import Units.Weapons.Weapon;
-import Units.Weapons.WeaponModel;
 
 /**
  * Commander Meridian focuses on a unique brand of flexibility; she transforms her tanks and artillery into each other 
@@ -70,8 +60,8 @@ public class Meridian extends Commander
     UnitModel arty = getUnitModel(UnitModel.UnitEnum.ARTILLERY);
     tank.moneyCostAdjustment      -= 500;
     arty.moneyCostAdjustment      += 500;
-    tank.possibleActions.add(new Transform(this, arty));
-    arty.possibleActions.add(new Transform(this, tank));
+    tank.possibleActions.add(new UnitActionType.Transform(arty));
+    arty.possibleActions.add(new UnitActionType.Transform(tank));
 
     addCommanderAbility(new ChangeAndFlow(this));
     addCommanderAbility(new VehicularCharge(this));
@@ -82,8 +72,15 @@ public class Meridian extends Commander
   {
     justTransformed.clear();
     toBeNerfed.clear();
-    modifyAbilityPower(40);
     return super.initTurn(map);
+  }
+
+  public void receiveUnitTranformEvent(Unit unit, UnitModel oldType)
+  {
+    if (this == unit.CO)
+    {
+      justTransformed.add(unit);
+    }
   }
 
   /**
@@ -187,138 +184,4 @@ public class Meridian extends Commander
   {
     return coInfo;
   }
-  
-  /** Effectively a wait, but the unit ends up as a different unit at the end of it. */
-  private static class Transform implements UnitActionType
-  {
-    public final UnitModel destinationType;
-    public final Meridian co;
-    
-    public Transform(Meridian user, UnitModel type)
-    {
-      co = user;
-      destinationType = type;
-    }
-    
-    @Override
-    public GameActionSet getPossibleActions(GameMap map, Path movePath, Unit actor)
-    {
-      XYCoord moveLocation = new XYCoord(movePath.getEnd().x, movePath.getEnd().y);
-      if( map.isLocationEmpty(actor, moveLocation) )
-      {
-        return new GameActionSet(new TransformAction(actor, movePath, this), false);
-      }
-      return null;
-    }
-
-    @Override
-    public String name()
-    {
-      return String.format("~%s", destinationType);
-    }
-  }
-
-  /** Effectively a wait, but the unit ends up as a different unit at the end of it. */
-  public static class TransformAction extends WaitAction
-  {
-    private Transform type;
-
-    public TransformAction(Unit unit, Path path, Transform pType)
-    {
-      super(unit, path);
-      type = pType;
-    }
-
-    @Override
-    public GameEventQueue getEvents(MapMaster gameMap)
-    {
-      GameEventQueue transformEvents = super.getEvents(gameMap);
-      
-      if( transformEvents.size() > 0 )
-      {
-        GameEvent moveEvent = transformEvents.peek();
-        if (moveEvent.getEndPoint().equals(waitLoc))
-        {
-          transformEvents.add(new UnitTransformEvent(actor, type.destinationType, type.co));
-        }
-      }
-      return transformEvents;
-    }
-    
-    @Override
-    public String toString()
-    {
-      return String.format("[Move %s to %s and transform to %s]", actor.toStringWithLocation(), waitLoc, type.destinationType);
-    }
-
-    @Override
-    public UnitActionType getUnitActionType()
-    {
-      return type;
-    }
-  }
-
-  /** Effectively a wait, but the unit ends up as a different unit at the end of it. */
-  public static class UnitTransformEvent implements GameEvent
-  {
-    private Unit unit;
-    private UnitModel type;
-    private Meridian user;
-
-    public UnitTransformEvent(Unit unit, UnitModel destination, Meridian listener)
-    {
-      this.unit = unit;
-      type = destination;
-      user = listener;
-    }
-
-    @Override
-    public GameAnimation getEventAnimation(MapView mapView)
-    {
-      return null;
-    }
-
-    @Override
-    public void sendToListener(GameEventListener listener)
-    {
-      // listener.receiveUnitDieEvent( this );
-    }
-
-    @Override
-    public void performEvent(MapMaster gameMap)
-    {
-      unit.model = type;
-      ArrayList<Weapon> temp = unit.weapons;
-      unit.weapons = new ArrayList<Weapon>();
-      
-      // Create the new weapon list
-      for( WeaponModel weapType : type.weaponModels )
-      {
-        unit.weapons.add(new Weapon(weapType));
-      }
-
-      // Try not to create ammo from nothing
-      for( int i = 0; i < temp.size() && i < unit.weapons.size(); i++ )
-      {
-        Weapon weap = unit.weapons.get(i);
-        if( !weap.model.hasInfiniteAmmo )
-          weap.ammo = temp.get(i).ammo;
-      }
-      
-      user.justTransformed.add(unit);
-    }
-
-    @Override
-    public XYCoord getStartPoint()
-    {
-      return new XYCoord(unit.x, unit.y);
-    }
-
-    @Override
-    public XYCoord getEndPoint()
-    {
-      return new XYCoord(unit.x, unit.y);
-    }
-  }
-
 }
